requirements.txt
tensorflow==1.15
Flask
numpy
Pillow
opencv-python
pandas
scikit-learn
tflearn